Responsive in progress, need to adapt the sidebar in responsive to finalize the responsive of all the pages
