+++ frontend/src/pages/HorseList.jsx
@@ -4,7 +4,7 @@ import UserContext from "@components/context/UserContext";
 import axios from "axios";
 import HorseCard from "@components/HorseCard";
 import Filters from "@components/Filters";
-import Header from "../components/SideBarMenu";
+import Header from "../components/SidebarMenu";
 import Banner from "../components/HeaderBannerHorses";
 
 function Home() {
@@ -53,13 +53,6 @@ function Home() {
         console.error(err);
       });
   };
-  const clearFilters = () => {
-    setStage0(true);
-    setStage1(true);
-    setStage2(true);
-    setStage3(true);
-    setStage4(true);
-  };
 
   useEffect(() => {
     getHorses();
@@ -72,22 +65,13 @@ function Home() {
         <Banner />
         <div className="rightContainerContent">
           <div className="horseListPage">
-            <div className="filtersBox">
-              <Filters
-                setStage0={setStage0}
-                setStage1={setStage1}
-                setStage2={setStage2}
-                setStage3={setStage3}
-                setStage4={setStage4}
-              />
-            </div>
-            <button
-              className="clearButton"
-              type="button"
-              onClick={() => clearFilters()}
-            >
-              Show all
-            </button>
+            <Filters
+              setStage0={setStage0}
+              setStage1={setStage1}
+              setStage2={setStage2}
+              setStage3={setStage3}
+              setStage4={setStage4}
+            />
             <div className="horseList">
               {stage2 &&
                 stage2Required.map((horse) => {
