+++ frontend/src/styles/components/listUser.scss
@@ -17,8 +17,8 @@
         max-height: 80px;
         border: 6px solid white;
         border-radius: 50%;
-        box-shadow: 4px 4px 20px rgba(56, 78, 119, 0.2);
-        transition: 0.3s cubic-bezier(0.6, 0.03, 0.28, 0.98); 
+        box-shadow: $shadowCard;
+        transition: 0.3s cubic-bezier(0.6, 0.03, 0.28, 0.98);
         transform: scale(1);
 
         &:hover {
@@ -26,9 +26,10 @@
           transform: scale(1.08);
         }
       }
+
       p {
         text-align: center;
-        opacity: 0.5;
+        color: $accentColor;
       }
     }
   }
